+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v1.2.0 (Upcomming)
+
+* Added Getter interface
+* Added the Get* and Has* utility functions (getter.go)
+* Added GetFromFirst(string, interface) (Contextvalue, Getter) function
+* Added DefaultGetter struct
+
 ## v1.1.0 (2014-06-06)
 
 * Added SyncContext function so Contexts are kept in sync via read/write mutex.